pyproject.toml
############################
# Project configuration
############################

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "my-langgraph-agent"
version = "0.1.0"
description = "A LangGraph-based agent service with FastAPI endpoints"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "httpx>=0.25.0",
  "python-dotenv>=1.0.0",
  # LangChain / LangGraph stack (adjust as needed)
  "langchain>=0.2.0",
  "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.5.0",
  "black>=24.3.0",
  "mypy>=1.8.0",
  "types-requests",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
exclude = [".venv", "venv", "build", "dist", "node_modules"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".venv", "venv", "build", "dist", "node_modules"]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.isort]
known-first-party = ["src"]
combine-as-imports = true

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
exclude = "(\\.venv|venv|build|dist|node_modules)"
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"


